extract MarketOption into its own component
